Completed Config Field Improvement
